+++ src/test/java/com/example/calculator/service/CalculationServiceTest.java
@@ -1,6 +1,5 @@
-package com.example.calculator.model;
+package com.example.calculator.service;
 
-import com.example.calculator.service.CalculationService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
